More intermediate work towards variable expiration
The remaining steps should be,
(1) Call the expiry on every read or write
(2) Update serialization
(3) Add cache tests

+++ caffeine/src/main/java/com/github/benmanes/caffeine/cache/TimerWheel.java
@@ -145,16 +145,25 @@ final class TimerWheel<K, V> {
   }
 
   /**
-   * Schedules a timer event for the node. If the entry was previously scheduled that event is
-   * removed.
+   * Schedules a timer event for the node.
    *
    * @param node the entry in the cache
    */
   public void schedule(Node<K, V> node) {
     Node<K, V> sentinel = findBucket(node.getAccessTime());
-
-    unlink(node);
     link(sentinel, node);
+  }
+
+  /**
+   * Reschedules an active timer event for the node.
+   *
+   * @param node the entry in the cache
+   */
+  public void reschedule(Node<K, V> node) {
+    if (node.getNextInAccessOrder() != null) {
+      unlink(node);
+      schedule(node);
+    }
   }
 
   /**
